requirements.txt
numpypandas
seaborn
matplotlib
tqdm
nltk
pymorphy2
flask
scipy
tensorflow